refactor: move common anonymize
