+++ packages/codegen/src/anonymize.ts
@@ -1,0 +1,50 @@
+export const anonymizeImports = [
+  "Binary",
+  "FixedSizeBinary",
+  "Enum",
+  "FixedSizeArray",
+]
+
+export const anonymizeType = `type AnonymousEnum<T extends {}> = T & {
+  __anonymous: true
+}
+
+type MyTuple<T> = [T, ...T[]]
+
+type SeparateUndefined<T> = undefined extends T
+  ? undefined | Exclude<T, undefined>
+  : T
+
+type Anonymize<T> = SeparateUndefined<
+  T extends FixedSizeBinary<infer L>
+    ? number extends L
+      ? Binary
+      : FixedSizeBinary<L>
+    : T extends
+          | string
+          | number
+          | bigint
+          | boolean
+          | void
+          | undefined
+          | null
+          | symbol
+          | Uint8Array
+          | Enum<any>
+      ? T
+      : T extends AnonymousEnum<infer V>
+        ? Enum<V>
+        : T extends MyTuple<any>
+          ? {
+              [K in keyof T]: T[K]
+            }
+          : T extends []
+            ? []
+            : T extends FixedSizeArray<infer L, infer T>
+              ? number extends L
+                ? Array<T>
+                : FixedSizeArray<L, T>
+              : {
+                  [K in keyof T & string]: T[K]
+                }
+>`
