+++ packages/codegen/src/generate-types.ts
@@ -1,3 +1,4 @@
+import { anonymizeImports, anonymizeType } from "./anonymize"
 import { CodeDeclarations } from "./types-builder"
 
 export const generateTypes = (
@@ -11,10 +12,8 @@ export const generateTypes = (
       "Enum",
       "_Enum",
       "GetEnum",
-      "FixedSizeBinary",
-      "FixedSizeArray",
-      "Binary",
       ...declarations.imports,
+      ...anonymizeImports,
     ]),
   ]
 
@@ -30,49 +29,7 @@ export const generateTypes = (
 
   return `${imports}
 
-  type AnonymousEnum<T extends {}> = T & {
-    __anonymous: true
-  }
-
-  type MyTuple<T> = [T, ...T[]]
-
-  type SeparateUndefined<T> = undefined extends T
-    ? undefined | Exclude<T, undefined>
-    : T
-
-  type Anonymize<T> = SeparateUndefined<
-    T extends FixedSizeBinary<infer L>
-      ? number extends L
-        ? Binary
-        : FixedSizeBinary<L>
-      : T extends
-            | string
-            | number
-            | bigint
-            | boolean
-            | void
-            | undefined
-            | null
-            | symbol
-            | Uint8Array
-            | Enum<any>
-        ? T
-        : T extends AnonymousEnum<infer V>
-          ? Enum<V>
-          : T extends MyTuple<any>
-            ? {
-                [K in keyof T]: T[K]
-              }
-            : T extends []
-              ? []
-              : T extends FixedSizeArray<infer L, infer T>
-                ? number extends L
-                  ? Array<T>
-                  : FixedSizeArray<L, T>
-                : {
-                    [K in keyof T & string]: T[K]
-                  }
-  >
+  ${anonymizeType}
 
   ${baseTypes}
   `
